requirements.txt
fastapi 
uvicorn 
pydantic 
orjson 
httpx
sqlalchemy 
sqlmodel 
psycopg 
pgvector
redis 
opentelemetry-api 
opentelemetry-sdk
python-dotenv 
rank-bm25 
sentence-transformers
transformers>=4.39.0
accelerate>=0.28.0
safetensors>=0.4.2
sentencepiece>=0.1.99
huggingface-hub>=0.20.3
onnxruntime 
tiktoken 
llama-cpp-python==0.2.*
networkx 
pytest 
pytest-asyncio 
pyyaml
